backend/requirements.txt
# API Framework
fastapi==0.109.0
uvicorn==0.27.0

# Database
sqlalchemy==2.0.25
pymongo==4.6.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7

# Environment and Utilities
python-dotenv==1.0.0
email-validator==2.1.0
pydantic==2.5.3
pydantic[email]

# HTTP
httpx==0.26.0

# Math and Science (for algorithm implementations)
numpy==1.26.3
scipy==1.12.0
sympy==1.12
matplotlib==3.8.2

# For testing
pytest==7.4.4